Refactored img2img workflow to load templates dynamically and inject advanced parameters for face restoration, tiling, hires fix, and refiner. Removed hardcoded workflow structure and integrated shared utility functions for better maintainability.
